[Fix] type 매칭 안되는 에러 해결

+++ src/main/java/com/chat_server/chatroom/dto/response/ChatRoomSummaryResponse.java
@@ -6,5 +6,5 @@ public record ChatRoomSummaryResponse(Long roomId,
                                       String chatType,
                                       String title,
                                       String profileUrl,
-                                      Long memberCount) {
+                                      Integer memberCount) {
 }
